requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
curl_cffi==0.7.4
groq==0.25.0
python-dotenv==1.1.0
pydantic==2.11.1
